+++ study-juc/src/main/java/org/ohx/studyjuc/semaphore/README.md
@@ -1,2 +1,11 @@
-# Semaphore
+# Semaphore 的学习 Demo
+
 单词Semaphore的中文含义是信号、信号系统。此类的主要作用就是限制线程并发的数量，如果不限制线程并发的数量，则CPU的资源很快就被耗尽，每个线程执行的任务是相当缓慢，因为CPU要把时间片分配给不同的线程对象，而且上下文切换也要耗时，最终造成系统运行效率大幅降低，所以限制并发线程的数量还是非常有必要的。
+
+## Demo1
+
+基本使用
+
+## addpermitscount
+
+本 Demo 说明许可数可以在运行时高于预设值。
