+++ study-juc/src/main/java/org/ohx/studyjuc/streamcommunication/README.md
@@ -1,1 +1,9 @@
-通过管道流进行线程间通信。
+# 通过管道流进行线程间通信。
+
+## pipeinputoutput
+
+通过管道字节流进行线程间通信。
+
+## pipereaderwriter
+
+通过管道字符流进行线程间通信。
