--- study-juc/src/main/java/org/ohx/studyjuc/streamcommunication/pipeinputoutput/README.md
@@ -1,1 +1,0 @@
-通过管道字节流进行线程间通信。
