--- study-juc/src/main/java/org/ohx/studyjuc/streamcommunication/pipereaderwriter/README.md
@@ -1,1 +1,0 @@
-通过管道字符流进行线程间通信。
