--- study-juc/src/main/java/org/ohx/studyjuc/suspendresume/suspendresumedeallock/README.md
@@ -1,1 +1,0 @@
-如果suspend()方法与resume()方法使用不当，极易造成公共同步对象被独占，其他线程无法访问公共同步对象的结果。
