--- study-juc/src/main/java/org/ohx/studyjuc/suspendresume/suspendresumenosamevalue/README.md
@@ -1,1 +1,0 @@
-在使用suspend()方法与resume()方法时也容易出现线程暂停，进而导致数据不完整的情况。
