--- study-juc/src/main/java/org/ohx/studyjuc/synchronizedtest/syntwolock/README.md
@@ -1,1 +1,0 @@
-验证 static synchronized 方法的锁 和 非static synchronized 方法不是同一个锁。
