+++ study-juc/src/main/java/org/ohx/studyjuc/threadlocal/README.md
@@ -1,1 +1,13 @@
-ThreadLocal的学习与使用
+# ThreadLocal 的学习 Demo
+
+## demo1
+
+单线程基本使用。
+
+## demo2
+
+多线程基本使用。
+
+## inheritablethreadlocaldemo
+
+ThreadLocal 是不具有继承性的，如果要实现子线程共享父线程的 ThreadLocal，需要使用 InheritableThreadLocal。
