+++ study-juc/src/main/java/org/ohx/studyjuc/threadlocal/inheritablethreadlocaldemo/Tools.java
@@ -1,4 +1,4 @@
-package org.ohx.studyjuc.threadlocal.inheritablethreadlocaltest1;
+package org.ohx.studyjuc.threadlocal.inheritablethreadlocaldemo;
 
 /**
  * @author mudkip
