+++ study-juc/src/main/java/org/ohx/studyjuc/threadpool/README.md
@@ -1,1 +1,10 @@
-Java线程池使用案例
+# 线程池学习 Demo
+
+## ThreadPoolExecutor
+
+`ThreadPoolExecutor`是 Java 线程池的实现。
+
+## Executors
+`Executors`是一个工具类，预先定义好了各种线程池。
+
+一般都不推荐直接使用，实际上都会根据业务去自定义线程池的各种参数。
